move YnabApi to its own file
